// src/app/+state/app.enum.ts
export enum DateFilterRange {
    Today = 'Today',
    Yesterday = 'Yesterday',
    SevenDays = 'Last 7 days',
    ThirtyDays = 'Last 30 days',
    ThisMonth = 'This month',
    CustomRange = 'Custom Range',
    CustomRangeApply = 'Apply',
}

export enum SortOrder {
    None = '',
    Asc = 'asc',
    Desc = 'desc'
}